+++ src/pyramid_bimt/views/__init__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 """Commonly shared view code."""
 
-from pyramid.security import has_permission
 from pyramid_bimt.static import app_assets
 from pyramid_bimt.static import form_assets
 from pyramid_deform import FormView as BaseFormView
+from colanderalchemy import SQLAlchemySchemaNode
 
 
 class FormView(BaseFormView):
@@ -21,24 +21,20 @@ class FormView(BaseFormView):
         the following:
          * requires static resources needed to render a form in a BIMT app;
          * sets ``self.title`` so the ``form.pt`` template can use it;
-         * removes from schema any nodes that have the ``edit_permission`` set,
-           if the current user does not have the permission set in
-           ``edit_permission``.
+         * if ``sqla_schema_class`` is set, it generates a Colander schema
+           from SQLAlchemy schema, using ColanderAlchemy.
 
         :returns: result from ``pyramid_deform.FormView.__call__()``
         """
         app_assets.need()
         form_assets.need()
-
-        self.schema.bind(request=self.request)
-        for child in self.schema.children:
-            if hasattr(child, 'edit_permission'):
-                if not has_permission(
-                    child.edit_permission,
-                    self.request.context,
-                    self.request,
-                ).boolval:
-                    self.schema.__delitem__(child.name)
+        if hasattr(self, 'sqla_schema_class') and self.sqla_schema_class:
+            fields = self.sqla_schema_class.allowed_fields(
+                self.request, mode='edit')
+            if not fields:
+                raise KeyError('No allowed fields found.')
+            self.schema = SQLAlchemySchemaNode(
+                self.sqla_schema_class, includes=fields)
 
         result = super(FormView, self).__call__()
         if isinstance(result, dict):  # pragma: no branch