+++ docs/CHANGELOG.rst
@@ -5,7 +5,11 @@ Changelog
 0.4.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Better support for ColanderAlchemy schemas in FormView.
+  [zupo]
+
+- Fix login URL in password reset email.
+  [zupo]
 
 
 0.4.4 (2013-12-31)